+++ tests/test_cli.py
@@ -127,7 +127,7 @@ def test_cli_extract_command_usage():
     result = runner.invoke(cli, ["extract"])
     assert result.exit_code > 0
     assert (
-        "Error: Missing option '-p' / '--parser'. Choose from:\n\tgelf,\n\tes"
+        "Error: Missing option '-p' / '--parser'. Choose from:\n\tgelf,\n\tes\n"
     ) in result.output
 
 
@@ -311,12 +311,17 @@ def test_cli_fetch_command_usage():
         "    --ldp-application-secret TEXT\n"
         "    --ldp-application-key TEXT\n"
         "    --ldp-endpoint TEXT\n"
+        "  mongo backend: \n"
+        "    --mongo-client-options KEY=VALUE,KEY=VALUE\n"
+        "    --mongo-collection TEXT\n"
+        "    --mongo-database TEXT\n"
+        "    --mongo-connection-uri TEXT\n"
         "  es backend: \n"
         "    --es-op-type TEXT\n"
         "    --es-client-options KEY=VALUE,KEY=VALUE\n"
         "    --es-index TEXT\n"
         "    --es-hosts TEXT\n"
-        "  -b, --backend [es|ldp|fs|swift|ws]\n"
+        "  -b, --backend [es|mongo|ldp|fs|swift|ws]\n"
         "                                  Backend  [required]\n"
         "  -c, --chunk-size INTEGER        Get events by chunks of size #\n"
     ) in result.output
@@ -325,7 +330,7 @@ def test_cli_fetch_command_usage():
     assert result.exit_code > 0
     assert (
         "Error: Missing option '-b' / '--backend'. "
-        "Choose from:\n\tes,\n\tldp,\n\tfs,\n\tswift,\n\tws\n"
+        "Choose from:\n\tes,\n\tmongo,\n\tldp,\n\tfs,\n\tswift,\n\tws\n"
     ) in result.output